refactor: FacadeService 도입
