+++ src/main/java/com/samhap/kokomen/interview/service/InterviewFacadeService.java
@@ -1,0 +1,94 @@
+package com.samhap.kokomen.interview.service;
+
+import com.samhap.kokomen.global.dto.ClientIp;
+import com.samhap.kokomen.global.dto.MemberAuth;
+import com.samhap.kokomen.global.exception.BadRequestException;
+import com.samhap.kokomen.global.service.RedisService;
+import com.samhap.kokomen.interview.domain.InterviewState;
+import com.samhap.kokomen.interview.domain.QuestionAndAnswers;
+import com.samhap.kokomen.interview.external.BedrockClient;
+import com.samhap.kokomen.interview.external.dto.response.InterviewSummaryResponses;
+import com.samhap.kokomen.interview.external.dto.response.LlmResponse;
+import com.samhap.kokomen.interview.service.dto.AnswerRequest;
+import com.samhap.kokomen.interview.service.dto.InterviewProceedResponse;
+import com.samhap.kokomen.interview.service.dto.InterviewRequest;
+import com.samhap.kokomen.interview.service.dto.InterviewResponse;
+import com.samhap.kokomen.interview.service.dto.InterviewResultResponse;
+import com.samhap.kokomen.interview.service.dto.InterviewStartResponse;
+import com.samhap.kokomen.interview.service.dto.InterviewSummaryResponse;
+import com.samhap.kokomen.member.service.MemberService;
+import java.time.Duration;
+import java.util.List;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+
+@RequiredArgsConstructor
+@Service
+public class InterviewFacadeService {
+
+    public static final String INTERVIEW_PROCEED_LOCK_KEY_PREFIX = "lock:interview:proceed:";
+
+    private final BedrockClient bedrockClient;
+    private final RedisService redisService;
+    private final InterviewService interviewService;
+    private final MemberService memberService;
+
+    public InterviewStartResponse startInterview(InterviewRequest interviewRequest, MemberAuth memberAuth) {
+        return interviewService.startInterview(interviewRequest, memberAuth);
+    }
+
+    public Optional<InterviewProceedResponse> proceedInterview(Long interviewId, Long curQuestionId, AnswerRequest answerRequest, MemberAuth memberAuth) {
+        memberService.validateHasToken(memberAuth);
+        interviewService.validateInterviewee(interviewId, memberAuth);
+        String lockKey = createInterviewProceedLockKey(memberAuth.memberId());
+        acquireLockForProceedInterview(lockKey);
+        try {
+            QuestionAndAnswers questionAndAnswers = interviewService.createQuestionAndAnswers(interviewId, curQuestionId, answerRequest, memberAuth);
+            LlmResponse llmResponse = bedrockClient.requestToBedrock(questionAndAnswers);
+            return interviewService.handleLlmResponse(memberAuth.memberId(), questionAndAnswers, llmResponse, interviewId);
+        } finally {
+            redisService.releaseLock(lockKey);
+        }
+    }
+
+    public static String createInterviewProceedLockKey(Long memberId) {
+        return INTERVIEW_PROCEED_LOCK_KEY_PREFIX + memberId;
+    }
+
+    private void acquireLockForProceedInterview(String lockKey) {
+        boolean lockAcquired = redisService.acquireLock(lockKey, Duration.ofSeconds(30));
+        if (!lockAcquired) {
+            throw new BadRequestException("이미 처리 중인 답변이 있습니다. 잠시 후 다시 시도해주세요.");
+        }
+    }
+
+    public void likeInterview(Long interviewId, MemberAuth memberAuth) {
+        interviewService.likeInterview(interviewId, memberAuth);
+    }
+
+    public InterviewResponse checkInterview(Long interviewId, MemberAuth memberAuth) {
+        return interviewService.checkInterview(interviewId, memberAuth);
+    }
+
+    public List<InterviewSummaryResponse> findMyInterviews(MemberAuth memberAuth, InterviewState state, Pageable pageable) {
+        return interviewService.findMyInterviews(memberAuth, state, pageable);
+    }
+
+    public InterviewSummaryResponses findOtherMemberInterviews(Long memberId, MemberAuth memberAuth, Pageable pageable) {
+        return interviewService.findOtherMemberInterviews(memberId, memberAuth, pageable);
+    }
+
+    public InterviewResultResponse findMyInterviewResult(Long interviewId, MemberAuth memberAuth) {
+        return interviewService.findMyInterviewResult(interviewId, memberAuth);
+    }
+
+    public InterviewResultResponse findOtherMemberInterviewResult(Long interviewId, MemberAuth memberAuth, ClientIp clientIp) {
+        return interviewService.findOtherMemberInterviewResult(interviewId, memberAuth, clientIp);
+    }
+
+    public void unlikeInterview(Long interviewId, MemberAuth memberAuth) {
+        interviewService.unlikeInterview(interviewId, memberAuth);
+    }
+}
